+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/MessagesScreen.kt
@@ -177,7 +177,7 @@ public fun MessagesScreen(
                 viewModel = listViewModel,
                 messagesLazyListState = rememberMessageListState(parentMessageId = currentState.parentMessageId),
                 onThreadClick = { message ->
-                    composerViewModel.setMessageMode(MessageMode.Thread(message))
+                    composerViewModel.setMessageMode(MessageMode.MessageThread(message))
                     listViewModel.openMessageThread(message)
                 },
                 onMediaGalleryPreviewResult = { result ->