+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/composer/MessageComposer.kt
@@ -379,7 +379,7 @@ public fun DefaultMessageComposerFooterContent(
     messageComposerState: MessageComposerState,
     onAlsoSendToChannelSelected: (Boolean) -> Unit,
 ) {
-    if (messageComposerState.messageMode is MessageMode.Thread) {
+    if (messageComposerState.messageMode is MessageMode.MessageThread) {
         Row(
             modifier = Modifier
                 .fillMaxWidth()