+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/header/MessageListHeader.kt
@@ -175,12 +175,12 @@ public fun DefaultMessageListHeaderCenterContent(
 ) {
     val title = when (messageMode) {
         MessageMode.Normal -> ChatTheme.channelNameFormatter.formatChannelName(channel, currentUser)
-        is MessageMode.Thread -> stringResource(id = R.string.stream_compose_thread_title)
+        is MessageMode.MessageThread -> stringResource(id = R.string.stream_compose_thread_title)
     }
 
     val subtitle = when (messageMode) {
         MessageMode.Normal -> channel.getMembersStatusText(LocalContext.current, currentUser)
-        is MessageMode.Thread -> stringResource(
+        is MessageMode.MessageThread -> stringResource(
             R.string.stream_compose_thread_subtitle,
             ChatTheme.channelNameFormatter.formatChannelName(channel, currentUser)
         )